+++ src/resources/v1/addons/index.ts
@@ -1,0 +1,26 @@
+// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
+
+export {
+  Addons,
+  type Addon,
+  type SetPackagePricing,
+  type SetPackagePricingResponse,
+  type AddonListResponse,
+  type AddonPublishResponse,
+  type AddonRemoveDraftResponse,
+  type AddonCreateParams,
+  type AddonUpdateParams,
+  type AddonListParams,
+  type AddonPublishParams,
+  type AddonSetPricingParams,
+  type AddonListResponsesMyCursorIDPage,
+} from './addons';
+export {
+  Entitlements,
+  type AddonPackageEntitlement,
+  type EntitlementCreateResponse,
+  type EntitlementListResponse,
+  type EntitlementCreateParams,
+  type EntitlementUpdateParams,
+  type EntitlementDeleteParams,
+} from './entitlements';